feat: started splitting framework
